test(runtime-class): cover the component element accessors from a browser fixture
