+++ packages/compiler/test/compile.test.js
@@ -11,6 +11,7 @@ import {
   getRuntimeEntryFiles,
   getRuntimeVersion,
 } from "@marko/compiler";
+import * as translator from "@marko/runtime-tags/translator";
 
 const template = path.join(
   import.meta.dirname,
@@ -94,6 +95,46 @@ describe("compiler/compile", () => {
           .id,
         /a%20b\.marko$/,
       );
+    });
+  });
+
+  // Analyze-phase mistakes are recorded as diagnostics rather than thrown, so
+  // one compile can report several; the fixture harness only ever reads
+  // `.message`, so nothing else exercises how the aggregate serializes.
+  describe("aggregate errors", () => {
+    const failing = (src) => {
+      try {
+        compileSync(src, "test.marko", { translator, code: false });
+      } catch (err) {
+        return err;
+      }
+      assert.fail("expected a compile error");
+    };
+
+    it("throws the error itself when there is only one", () => {
+      const err = failing("<if>a</if>");
+      assert.equal(err.name, "CompileError");
+      assert.equal(err.errors, undefined);
+    });
+
+    it("collects several into one CompileErrors", () => {
+      const err = failing("<if>a</if>\n<if>b</if>");
+      assert.equal(err.name, "CompileErrors");
+      assert.equal(err.errors.length, 2);
+    });
+
+    it("names itself and strips ansi when stringified", () => {
+      const text = String(failing("<if>a</if>\n<if>b</if>"));
+      assert.ok(text.startsWith("CompileErrors: "), text);
+      assert.doesNotMatch(text, /\u001b\[/, "expected no ansi escapes");
+      assert.match(text, /test\.marko:1/);
+      assert.match(text, /test\.marko:2/);
+    });
+
+    it("serializes to the text a bundler would log", () => {
+      const err = failing("<if>a</if>\n<if>b</if>");
+      assert.equal(err.toJSON(), String(err));
+      assert.equal(JSON.parse(JSON.stringify({ err })).err, String(err));
     });
   });
